FROMLIST: sched/fair: Generalize misfit lb by adding a misfit reason
MISFIT_PERF is what is currently implemented. It indicates that the task
requires moving to a more performant CPU.

Guard misfit handling in find_busiest_queue and update_sd_pick_busiest
with MISFIT_PERF. They explicitly assume this type of misfit

This generalizes misfit lb to allow for more types of misfits to be
added. Like MISFIT_POWER to help uclamp_max being more effective, and
MISFIT_LATENCY to help latency sensitive tasks to be spread in
oversubscribe conditions.

Bug: 340190651
Link: https://lore.kernel.org/lkml/20231209011759.398021-3-qyousef@layalina.io/
Change-Id: Id798e8936dddcf368c2956bad708e99087c1076f
Signed-off-by: Qais Yousef <qyousef@layalina.io>
Signed-off-by: Qais Yousef <qyousef@google.com>
[jstultz: Added lore link, small build fix from bad cherry-pick,
 whitespace indention fixup]
Signed-off-by: John Stultz <jstultz@google.com>