ANDROID: eevdf: Add vh for update_deadline()
In kernel6.6+, the EEVDF has replaced CFS, and eevdf allows tasks
to run for any amount of time. Even if a task only runs for
a very small period of time, it may be preempted.

Sometimes we need some specific threads, such as lock owner, UI and
other threads, to be able to run for at least a specific time
before being preempted to improve fluency.

So add update_deadline's vh to allow vendor can contol the specific tasks.

Bug: 320209712
Change-Id: I06200fb2fff810624cf802daf39796ed2466f91f
Signed-off-by: Xuewen Yan <xuewen.yan@unisoc.com>